refactor(cubeops): extract service layer from handlers
